[meta] General styling cleanup and data type alignment

+++ src/ddraw/ddraw_common_surface.h
@@ -181,7 +181,7 @@ namespace dxvk {
       return static_cast<float>(input) / static_cast<float>(max);
     }
 
-    uint16_t GetMipCount() const {
+    uint32_t GetMipCount() const {
       return m_mipCount;
     }
 
@@ -582,28 +582,28 @@ namespace dxvk {
       return ColorKeyToARGB(pixelFormat, colorKey != nullptr ? colorKey->dwColorSpaceLowValue : 0u);
     }
 
-    bool                             m_dirtyDDraw         = false;
-    bool                             m_dirtyD3D9          = false;
+    bool                             m_dirtyDDraw       = false;
+    bool                             m_dirtyD3D9        = false;
 
-    bool                             m_isDesc2Set         = false;
-    bool                             m_isDescSet          = false;
-    bool                             m_hasTextureHandle   = false;
+    bool                             m_isDesc2Set       = false;
+    bool                             m_isDescSet        = false;
+    bool                             m_hasTextureHandle = false;
 
-    bool                             m_isAttached         = false;
-    bool                             m_isRenderTarget     = false;
+    bool                             m_isAttached       = false;
+    bool                             m_isRenderTarget   = false;
     bool                             m_isBackBufferOrFlippable = false;
 
     Com<DDrawCommonInterface>        m_commonIntf;
 
-    D3DCommonDevice*                 m_commonD3DDevice    = nullptr;
+    D3DCommonDevice*                 m_commonD3DDevice  = nullptr;
 
-    uint16_t                         m_mipCount           = 1;
-    uint32_t                         m_backBufferIndex    = 0;
+    uint32_t                         m_mipCount         = 1;
+    uint32_t                         m_backBufferIndex  = 0;
 
-    DDSURFACEDESC                    m_desc               = { };
-    DDSURFACEDESC2                   m_desc2              = { };
-    D3D9SurfaceType                  m_d3d9SurfaceType    = D3D9SurfaceType::None;
-    RECT                             m_rect               = { };
+    DDSURFACEDESC                    m_desc             = { };
+    DDSURFACEDESC2                   m_desc2            = { };
+    D3D9SurfaceType                  m_d3d9SurfaceType  = D3D9SurfaceType::None;
+    RECT                             m_rect             = { };
 
     Com<DDrawClipper>                m_clipper;
     Com<DDrawPalette>                m_palette;
@@ -612,19 +612,19 @@ namespace dxvk {
     Com<d3d9::IDirect3DTexture9>     m_texture9;
     Com<d3d9::IDirect3DCubeTexture9> m_cubeMap9;
 
-    d3d9::D3DFORMAT                  m_format9            = d3d9::D3DFMT_UNKNOWN;
+    d3d9::D3DFORMAT                  m_format9          = d3d9::D3DFMT_UNKNOWN;
 
-    DDCOLORKEY                       m_ckNormalized       = { };
+    DDCOLORKEY                       m_ckNormalized     = { };
 
-    DDraw7Surface*                   m_surf7              = nullptr;
-    DDraw4Surface*                   m_surf4              = nullptr;
-    DDraw3Surface*                   m_surf3              = nullptr;
-    DDraw2Surface*                   m_surf2              = nullptr;
-    DDrawSurface*                    m_surf               = nullptr;
+    DDraw7Surface*                   m_surf7            = nullptr;
+    DDraw4Surface*                   m_surf4            = nullptr;
+    DDraw3Surface*                   m_surf3            = nullptr;
+    DDraw2Surface*                   m_surf2            = nullptr;
+    DDrawSurface*                    m_surf             = nullptr;
 
     // Track the origin surface, as in the DDraw surface
     // that gets created through a CreateSurface call
-    IUnknown*                        m_origin             = nullptr;
+    IUnknown*                        m_origin           = nullptr;
 
   };
